Implemented - Multiple Spawning + DestroyBlock Script
This commit has blocks spawn in quick succession, in a way that makes it look sort of on purpose, and also has them removed from the scene when they get to -150 on the Z as this prevents the scene from getting too cluttered (as well as preventing lag from happening).

Next, implement a way to keep track of how "far" the player is going in meters.