Adjusted/Implemented - Block Spawn Rate/Pos + Fog Material
This commit adjusts the spawn rate of the blocks (still might need to be slightly quicker) and also changes where they start spawning from. A fog material has also been added, it's not quite like I want it right now but it's fine for now.

Next, implement collision for the player and the blocks.