Adjusted/Added - Block Density + New Block Variants
This commit first adjusts the block density so that more appear within the player's view (while also keeping the same speed/spawn rate). After this a raycast was added to prevent blocks from spawning directly on-top of other blocks, and finally 5 new block variants have been added that each appear the further a player gets into a run.

A cool two material specical block appears as the 10th (and currently final) block variant. Not many players should be able to reach this just now.

Next, add a start screen similar to that of CubeField that has the game continusouly running in the background until it starts.