Fixed/Changed - No Movement Collision + To Mesh Collider
This commit fixes the issue with the player being able to avoid colliding with blocks if they don't move at all. It also switches from a box collider to a mesh collider on the player, this prevents the player from colliding with blocks when they feel they shouldn't have as the collider actually fits the mesh.

Next, adjust the camera to lag less as this was a complaint we had from play-testing.