Fixed/Implemented - Different Block Speed + Block Fade In
This commit fixes the issue with blocks moving at different speeds when  the game speeds up every 500 meters. It also implements a fade in effect for the blocks which gives them a better feel and also gives the player a bit of time to adjust to the new spawned block.

Next, adjust how close the blocks can spawn to the player to be further away.