Improved - Player Movement + Camera
This commit works on improving the overall feel of the player movement. Before the moving the player would begin to feel clunky at around the 5000m mark and even before that moving from one direction to the other felt off, that has all now been addressed.

Next, fix the issue with the renderer not being found for block 10.