Improved/Implemented - Grass/Snow Materials + Better Spawning
This commit first improves the look of the Grass and Snow zones then improves the block spawning by decreasing the spawnInterval the further the player gets into a run. This prevents the block spawning from getting really thin, leaving only 1 or 2 blocks to avoid.

Next, build & package for v0.1.3 release.